more changes to 64-bit safe API
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@16056 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKEditor.m
@@ -1897,7 +1897,7 @@ enum { BDSKMoveToTrashAsk = -1, BDSKMoveToTrashNo = 0, BDSKMoveToTrashYes = 1 };
     
     if (nil == error) {
         // shouldn't get here
-        NSLog(@"%@:%d formatter failed for unknown reason", __FILENAMEASNSSTRING__, __LINE__);
+        NSLog(@"%@:%ld formatter failed for unknown reason", __FILENAMEASNSSTRING__, (long)__LINE__);
     } else if (control == tableView) {
         
         NSString *fieldName = [fields objectAtIndex:[tableView editedRow]];
